refactor: move lstm model code from train/inference to mode/lstm.py
