+++ modeling/src/inference.py
@@ -7,6 +7,7 @@ import numpy as np
 import torch
 import torch.nn as nn
 from sklearn.preprocessing import MinMaxScaler
+from model.lstm import MultiOutputLSTM
 
 WINDOW_SIZE = 30
 
@@ -33,16 +34,6 @@ def get_scalers():
     outputs_temperature, outputs_PM = get_outputs()
     return get_scaler_temperature(outputs_temperature), get_scaler_PM(outputs_PM)
 
-class MultiOutputLSTM(nn.Module):
-    def __init__(self, outputs, input_size=3, hidden_size=64, num_layers=1):
-        super().__init__()
-        self.lstm = nn.LSTM(input_size, hidden_size, num_layers, batch_first=True)
-        self.fc = nn.Linear(hidden_size, len(outputs))
-    
-    def forward(self, x):
-        out, _ = self.lstm(x)
-        return self.fc(out[:, -1, :])
-    
 def init_model():
     outputs_temperature, outputs_PM = get_outputs()
