+++ modeling/src/model/lstm.py
@@ -1,0 +1,12 @@
+import torch
+import torch.nn as nn
+
+class MultiOutputLSTM(nn.Module):
+    def __init__(self, outputs, input_size=3, hidden_size=64, num_layers=1):
+        super().__init__()
+        self.lstm = nn.LSTM(input_size, hidden_size, num_layers, batch_first=True)
+        self.fc = nn.Linear(hidden_size, len(outputs))
+    
+    def forward(self, x):
+        out, _ = self.lstm(x)
+        return self.fc(out[:, -1, :])
